+++ mnelab/utils/__init__.py
@@ -3,5 +3,5 @@
 # License: BSD (3-clause)
 
 from .dependencies import have
-from .xdf import parse_xdf, parse_chunks, read_raw_xdf
-from .utils import split_fname
+from .io import (IMPORT_FORMATS, EXPORT_FORMATS, split_fname, parse_xdf,
+                 parse_chunks, read_raw_xdf)
